scripts/requirements.txt
httpx
google-cloud-storage